requirements.txt
Flask>=3.1.2
Pillow>=11.3.0
requests>=2.32.5
